--- src/app/component/hooks/UseAuth.tsx
@@ -1,4 +1,0 @@
-"use client";
-import { useState, useEffect, createContext, useContext } from "react";
-
-interface AutContexType {}
