+++ src/app/component/navbar/NavbarHome.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 import profil from "../../../../public/asset/porfil.png";
 import { Search } from "lucide-react";
 import Link from "next/link";
-import { useHook } from "@/app/layout/Provider";
+import { useHook } from "../hooks/UserAuth";
 
 const NavbarHome: React.FC = () => {
   const { currentUser, isLoading } = useHook();
