+++ src/app/component/navbar/NavbarProfil.tsx
@@ -2,7 +2,7 @@ import iconHItam from "../../../../public/asset/IconHitam.png";
 import Image from "next/image";
 import profil from "../../../../public/asset/porfil.png";
 import Link from "next/link";
-import { useHook } from "@/app/layout/Provider";
+import { useHook } from "../hooks/UserAuth";
 
 const NavbarProfil: React.FC = () => {
   const { currentUser, isLoading } = useHook();
